Refactor : #16 select, write, layout 페이지 전체 병합
