+++ src/pages/App.tsx
@@ -7,13 +7,12 @@ import theme from '../common/style/theme';
 import LandingPage from '.';
 import Main from './Main';
 import GuestHome from './GuestHome';
-import LeafSelect from './LeafSelect';
 import Layout from '../components/Layout/Layout';
 import AnniversaryManagement from './AnniversaryManagement';
 import LoginPage from './LoginPage';
 import NicknameRegist from './NicknameRegist';
 import SocialRedirect from './SocialLogin/SocialRedirect';
-import GuestWrite from './GuestWrite';
+import GuestFlow from './GuestFlow';
 
 function App() {
   return (
@@ -30,12 +29,7 @@ function App() {
             <Route path="/login/redirect" element={<SocialRedirect />} />
             <Route path="/login/nickname" element={<NicknameRegist />} />
             <Route path="/:postId" element={<GuestHome />} />
-            {/* <Route path="/guest/write" element={<GuestWrite />} /> */}
-            <Route path="/guest/select/:postId" element={<LeafSelect />} />
-            <Route
-              path="/guest/write/:postId/:iconId"
-              element={<GuestWrite />}
-            />
+            <Route path="/guest/:postId" element={<GuestFlow />} />
           </Routes>
         </Layout>
       </Router>
